handling movement to the right in a logically correct way. See notes
No graphics yet. The Server receives an action request correctly from the Client, then checks if the movement is correct (which it is, in all trials yet tested) and executes the proper action logically. After that, the server resends the message to the client, who receives it correctly. Now the graphics should be handled for this case